+++ packages/core/src/connectivity/scp-cf/index.ts
@@ -1,6 +1,7 @@
 export * from './jwt';
 export * from './cache';
 export * from './legacy/client-credentials-token';
+export * from './legacy/xsuaa-service';
 export * from './client-credentials-token-cache';
 export * from './connectivity-service';
 export * from './connectivity-service-types';
@@ -9,7 +10,7 @@ export * from './environment-accessor';
 export * from './environment-accessor-types';
 export * from './resilience-options';
 export * from './token-accessor';
-export * from './legacy/xsuaa-service';
+export * from './verification-keys';
 export * from './xsuaa-service-types';
 export * from './protocol';
 export * from './authorization-header';
